gmail-event-processor/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-cloud-secret-manager
google-cloud-firestore
google-generativeai
beautifulsoup4
